+++ packages/python/qsu/math/__init__.py
@@ -1,6 +1,7 @@
 from .ceil import ceil
 from .clamp import clamp
 from .div import div
+from .floor import floor
 from .mul import mul
 from .numPick import numPick
 from .numUnique import numUnique
@@ -12,6 +13,7 @@ __all__ = [
 	'ceil',
 	'clamp',
 	'div',
+	'floor',
 	'mul',
 	'numPick',
 	'numUnique',